// backend/src/enums/UserEnums.ts
export enum Role {
    STUDENT = "student",
    LECTURER = "lecturer",
    ADMIN = "admin"
  }
  
  export enum StudentType {
    STUDENT = "student",
    REP = "rep"
  }
  
  export enum LecturerType {
    LECTURER = "lecturer",
    HOD = "hod",
    TRAINEE = "trainee"
  }
  
  export enum AdminType {
    ACADEMIC = "academic",
    RESOURCE = "resource",
    EVENT = "event"
  }
  
  export enum Gender {
    MALE = "male",
    FEMALE = "female",
    OTHER = "other"
  }